+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- entity type selection triggers search in merged view
+
 ### Changes
+
+- resolve identifiers in merge view
 
 ### Deprecated
 
@@ -18,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - fix editing for link fields without a title value
+- only show allowed precisions in temporal field drop-down
 
 ### Security